+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/domain/usecase/param/SignInParams.kt
@@ -1,4 +1,4 @@
-package com.gdavidpb.tuindice.login.domain.param
+package com.gdavidpb.tuindice.login.domain.usecase.param
 
 data class SignInParams(
 	val usbId: String,
