+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/domain/usecase/validator/SignInParamsValidator.kt
@@ -3,7 +3,7 @@ package com.gdavidpb.tuindice.login.domain.usecase.validator
 import com.gdavidpb.tuindice.base.domain.usecase.base.ParamsValidator
 import com.gdavidpb.tuindice.login.domain.usecase.error.SignInError
 import com.gdavidpb.tuindice.login.domain.exception.SignInIllegalArgumentException
-import com.gdavidpb.tuindice.login.domain.param.SignInParams
+import com.gdavidpb.tuindice.login.domain.usecase.param.SignInParams
 import com.gdavidpb.tuindice.login.utils.extension.isUsbId
 
 class SignInParamsValidator : ParamsValidator<SignInParams> {
