+++ login/src/main/kotlin/com/gdavidpb/tuindice/login/presentation/viewmodel/SignInViewModel.kt
@@ -1,7 +1,7 @@
 package com.gdavidpb.tuindice.login.presentation.viewmodel
 
 import com.gdavidpb.tuindice.base.presentation.viewmodel.BaseViewModel
-import com.gdavidpb.tuindice.login.domain.param.SignInParams
+import com.gdavidpb.tuindice.login.domain.usecase.param.SignInParams
 import com.gdavidpb.tuindice.login.presentation.contract.SignIn
 import com.gdavidpb.tuindice.login.presentation.reducer.SignInReducer
